refactor: separate forms and models into different files
